+++ noodles-bam/examples/bam_write_async.rs
@@ -8,7 +8,10 @@ use noodles_bam as bam;
 use noodles_sam::{
     self as sam,
     alignment::Record,
-    header::{Program, ReferenceSequence},
+    header::{
+        record::value::{map::Program, Map},
+        ReferenceSequence,
+    },
 };
 use tokio::io;
 
@@ -19,7 +22,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let header = sam::Header::builder()
         .set_header(Default::default())
         .add_reference_sequence(ReferenceSequence::new("sq0".parse()?, 8)?)
-        .add_program(Program::new("noodles-bam"))
+        .add_program(Map::<Program>::new("noodles-bam"))
         .add_comment("an example BAM written by noodles-bam")
         .build();