+++ noodles-sam/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Changed
 
-  * sam/header: Rewrap `Header` as a record map value (`Map`).
+  * sam/header: Rewrap `header::Header` and `Program` as a record map value
+    (`Map`).
 
 ## 0.18.0 - 2022-08-16